Created Empty Files
Created blank html document with a empty script tag, called index.html.
This file will contain the html content and javascript.

Created a blank document (styles.css). This file will contain the css
styling for the application.

Created an empty text document (algorithm.txt). This file is to write
out pseudocode for the application - to eventually be converted to
javascript in the html file.