HTML, CSS, Images
Whelp I'm still getting use to committing often. Here's the changes I've
made:

HTML:
Added HTML content for the webpage - Header, score buttons, images
(hidden) to display both the user and cpu selection, buttons for the
user to make their selection.

CSS:
Adding styling for all the HTML content that was added.

Images:
Added 3 images (rock, paper and scissors). To display the user and cpu
decisions for a given round. For the user to make their selection using
the buttons. Included links to the images in the README.md file, will
need to add full credit later, but these images are just placeholders
for now.