Added Popups
Added play and play again popups to the RPS game. Styled the popups to
be identical. Set the display to none for the play again popup. This
popup will be displayed at the conclusion of a game. The play popup's
display will be set to none (making it invisible) once the play button
is hit.