Updated Algorithm to Reduce the Amount of Code
Changed the algorithm in algorithm.txt. Added a choose variable
to store all the previous choose buttons (chooseRock, choosePaper,
chooseScissors). Used a single event listener to listen for any of these
buttons being clicked. Changed the playRound() function to accept the
event as a parameter and execute the other functions instead of calling
them one by one in the event listener function.